To save memory ignore lsp minor mode in desktop restore
This prevents the lsp minor modes from being saved in the desktop
state, so that the modes are not restored on desktop restore.

Without this, if many different projects are open, then restoring
desktop fires up all the associated language servers which hog
memory.

Now we can ensure a server starts only if we visit a project, by
hooking into lsp-deferred for the given language mode.